Hemera: Implement factory reset
The user must put a file in the directory /usr/share/hemera/factoryreset.d/
with the name of the device replacing slashes with underscores (e.g.
/usr/share/hemera/factoryreset.d/dev_sda2).
Every line of the file must contain the commands to execute in case of
factory reset. It is possible to insert comments using # at the
beginning of a line.
If the factory reset fails, fsck continues.
TODO: do something if factory reset fails.